Restructure: Decouple web from ginclient
Web package does not depend on ginclient.  User token managements moved
to the ginclient package.
